requirements.txt
pytest==7.1.2
selene>=2.0.b5
allure-pytest